requirements.txt
playwright
azure-cognitiveservices-speech
openai
moviepy
python-pptx
nbformat
pandas
python-dotenv